Prevent ties by increasing the votes against the loser.
Subject 25 interpreted the occasion of a tie as proof that the coplayers were fake, forgetting, I guess, the earlier statement that ties are broken at random. I hate when subjects are too smart to get tricked but too dumb to realize a flaw in their reasoning why they were tricked.